Use the right type conversion
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@1463783 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/core/ActiveMQConnectionFactory.cpp
@@ -207,11 +207,11 @@ namespace core{
             this->sendAcksAsync = Boolean::parseBoolean(
                 properties->getProperty("connection.sendAcksAsync", Boolean::toString(sendAcksAsync)));
             this->optimizeAcknowledgeTimeOut = Long::parseLong(
-                properties->getProperty("connection.optimizeAcknowledgeTimeOut", Integer::toString(optimizeAcknowledgeTimeOut)));
+                properties->getProperty("connection.optimizeAcknowledgeTimeOut", Long::toString(optimizeAcknowledgeTimeOut)));
             this->optimizedAckScheduledAckInterval = Long::parseLong(
-                properties->getProperty("connection.optimizedAckScheduledAckInterval", Integer::toString(optimizedAckScheduledAckInterval)));
+                properties->getProperty("connection.optimizedAckScheduledAckInterval", Long::toString(optimizedAckScheduledAckInterval)));
             this->consumerFailoverRedeliveryWaitPeriod = Long::parseLong(
-                properties->getProperty("connection.consumerFailoverRedeliveryWaitPeriod", Integer::toString(consumerFailoverRedeliveryWaitPeriod)));
+                properties->getProperty("connection.consumerFailoverRedeliveryWaitPeriod", Long::toString(consumerFailoverRedeliveryWaitPeriod)));
 
             this->defaultPrefetchPolicy->configure(*properties);
             this->defaultRedeliveryPolicy->configure(*properties);
